refactor(settings): fix circular dependency between ui and data
