+++ libs/settings/data/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/tokens';
+export { SettingsApi } from './lib/settings-api';
