+++ libs/settings/public/interfaces/src/index.ts
@@ -1,3 +1,3 @@
-export { SettingsApi } from './lib/settings-api';
+export { SettingsApi } from '@architect-poc/settings/data';
 export { SettingsType } from '@architect-poc/settings/domain';
 export { SettingsInterfacesModule } from './lib/settings-interfaces.module';
